moved location of an answer write from behaviour to component folder
